+++ test/dummy/app/models/fake_block.rb
@@ -3,5 +3,7 @@ class FakeBlock < ApplicationRecord
                               root: false,
                               allowed_parent_types: ["RecordingStudioPresskits::PressKit"]
 
+  include RecordingStudio::Capabilities::Trashable.to
+
   validates :title, presence: true
 end