+++ test/dummy/app/models/workspace.rb
@@ -1,4 +1,6 @@
 class Workspace < ApplicationRecord
   recording_studio_recordable label: "Workspace", root: true
-  RecordingStudio.enable_capability(:accessible, on: self) if defined?(RecordingStudioAccessible)
+  RecordingStudio.enable_capability(:accessible, on: self)
+
+  include RecordingStudio::Capabilities::Orderable.to(allows: ["RecordingStudioPresskits::PressKit"])
 end